Server/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
backports.tarfile==1.2.0
Bottleneck==1.3.7
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
exceptiongroup==1.2.0
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.52.4
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
google-search-results==2.4.2
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.3
keras==3.3.3
keyring==25.2.1
kiwisolver==1.4.5
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
libclang==18.1.1
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
nh3==0.2.17
numexpr==2.8.7
numpy==1.26.4
openai>=1.50.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
Pillow==10.0.0
pinecone-client==2.2.4
pip==23.3.1
pkginfo==1.10.0
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pyloco==0.0.139
pyparsing==3.1.2
pyshine==0.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
readme_renderer==43.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
safetensors==0.4.2
serpapi==0.1.5
setuptools==68.2.2
SimpleWebSocketServer==0.1.2
six==1.16.0
sniffio==1.3.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.3.0
tokenizers==0.15.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.39.3
twine==5.1.0
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
ushlex==0.99.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.41.2
wrapt==1.16.0
zipp==3.19.0
